\name{rma.sd.assessment}
\docType{data}
\alias{rma.sd.assessment}
\title{An example of the result of an SD assessment}
\description{
  The Dilution files were processed with the
  \code{affy} version 1.0 package rma add-on function,
  and then the function \code{\link[affycomp]{assessSD}}
  from the affycomp package was applied.
}
\usage{data(rma.sd.assessment)}
\format{A list.}


\keyword{datasets}
