\name{rma.assessment}
\docType{data}
\alias{rma.assessment}
\alias{rma.assessment.133}
\alias{rma.assessment2}
\alias{rma.assessment2.133}

\title{An example of the result of an assessment}
\description{
  The Dilution and Spike-in cel files were processed with the
  \code{affy} version 1.0 package rma add-on function,
  and then the functions \code{\link[affycomp]{assessAll}} and
  \code{\link[affycomp]{assessSpikeIn2}} from the affycomp
  package were applied. 
}
\usage{data(rma.assessment)
data(rma.assessment.133)
data(rma.assessment2)
data(rma.assessment2.133)
}
\format{A list of list.}


\keyword{datasets}
